+++ Sources/BlipCore/BlipCore.swift
@@ -3,5 +3,5 @@
 import Foundation
 
 public enum BlipCore {
-    public static let version = "0.3.0"
+    public static let version = "0.3.1"
 }